optimize space for in-place quicksort

+++ src/main.rs
@@ -133,9 +133,8 @@ fn main() {
         // Quick_sort
         for d in map_files.iter() {
             println!("Running qs file: {}", d);
-            let mut v: Vec<i32> = read_csv::<i32>(&d);
-            // let v_r: Vec<&Vec<u128>> = v.iter().map(|f| f).collect();
-            let res = run_quick_sort_benchmark(d, &mut v, |a: &i32, b: &i32| -> i32 {*a -*b}, rounds, tn);
+            let mut v: Vec<i16> = read_csv::<i16>(&d);
+            let res = run_quick_sort_benchmark(d, &mut v, |a: &i16, b: &i16| -> i32 { *a - *b } as i32, rounds, tn);
             qs_res.extend(res);
         }
         println!("Writing qs result");
